simplify changeKeys
